ARROW-12751: [C++] Move element_wise_min/max to scalar_compare.cc
