+++ cpp/src/arrow/compute/kernels/scalar_arithmetic.cc
@@ -18,10 +18,8 @@
 #include <cmath>
 #include <limits>
 
-#include "arrow/compute/api_scalar.h"
 #include "arrow/compute/kernels/common.h"
 #include "arrow/type_traits.h"
-#include "arrow/util/bitmap_ops.h"
 #include "arrow/util/int_util_internal.h"
 #include "arrow/util/macros.h"
 
@@ -399,60 +397,6 @@ struct PowerChecked {
   }
 };
 
-struct Minimum {
-  template <typename T>
-  static enable_if_floating_point<T> Call(T left, T right) {
-    return std::fmin(left, right);
-  }
-
-  template <typename T>
-  static enable_if_integer<T> Call(T left, T right) {
-    return std::min(left, right);
-  }
-
-  template <typename T>
-  static constexpr enable_if_t<std::is_same<float, T>::value, T> antiextreme() {
-    return std::nanf("");
-  }
-
-  template <typename T>
-  static constexpr enable_if_t<std::is_same<double, T>::value, T> antiextreme() {
-    return std::nan("");
-  }
-
-  template <typename T>
-  static constexpr enable_if_integer<T> antiextreme() {
-    return std::numeric_limits<T>::max();
-  }
-};
-
-struct Maximum {
-  template <typename T>
-  static enable_if_floating_point<T> Call(T left, T right) {
-    return std::fmax(left, right);
-  }
-
-  template <typename T>
-  static enable_if_integer<T> Call(T left, T right) {
-    return std::max(left, right);
-  }
-
-  template <typename T>
-  static constexpr enable_if_t<std::is_same<float, T>::value, T> antiextreme() {
-    return std::nanf("");
-  }
-
-  template <typename T>
-  static constexpr enable_if_t<std::is_same<double, T>::value, T> antiextreme() {
-    return std::nan("");
-  }
-
-  template <typename T>
-  static constexpr enable_if_integer<T> antiextreme() {
-    return std::numeric_limits<T>::min();
-  }
-};
-
 // Generate a kernel given an arithmetic functor
 template <template <typename... Args> class KernelGenerator, typename Op>
 ArrayKernelExec ArithmeticExecFromOp(detail::GetTypeId get_id) {
@@ -502,28 +446,6 @@ struct ArithmeticFunction : ScalarFunction {
       if (auto type = CommonNumeric(*values)) {
         ReplaceTypes(type, values);
       }
-    }
-
-    if (auto kernel = DispatchExactImpl(this, *values)) return kernel;
-    return arrow::compute::detail::NoMatchingKernel(this, *values);
-  }
-};
-
-struct ArithmeticVarArgsFunction : ScalarFunction {
-  using ScalarFunction::ScalarFunction;
-
-  Result<const Kernel*> DispatchBest(std::vector<ValueDescr>* values) const override {
-    RETURN_NOT_OK(CheckArity(*values));
-
-    using arrow::compute::detail::DispatchExactImpl;
-    if (auto kernel = DispatchExactImpl(this, *values)) return kernel;
-
-    EnsureDictionaryDecoded(values);
-
-    if (auto type = CommonNumeric(*values)) {
-      ReplaceTypes(type, values);
-    } else if (auto type = CommonTimestamp(*values)) {
-      ReplaceTypes(type, values);
     }
 
     if (auto kernel = DispatchExactImpl(this, *values)) return kernel;
@@ -590,176 +512,6 @@ std::shared_ptr<ScalarFunction> MakeUnarySignedArithmeticFunctionNotNull(
       auto exec = ArithmeticExecFromOp<ScalarUnaryNotNull, Op>(ty);
       DCHECK_OK(func->AddKernel({ty}, ty, exec));
     }
-  }
-  return func;
-}
-
-using MinMaxState = OptionsWrapper<ElementWiseAggregateOptions>;
-
-// Implement a variadic scalar min/max kernel.
-template <typename OutType, typename Op>
-struct ScalarMinMax {
-  using OutValue = typename GetOutputType<OutType>::T;
-
-  static void ExecScalar(const ExecBatch& batch,
-                         const ElementWiseAggregateOptions& options, Scalar* out) {
-    // All arguments are scalar
-    OutValue value{};
-    bool valid = false;
-    for (const auto& arg : batch.values) {
-      // Ignore non-scalar arguments so we can use it in the mixed-scalar-and-array case
-      if (!arg.is_scalar()) continue;
-      const auto& scalar = *arg.scalar();
-      if (!scalar.is_valid) {
-        if (options.skip_nulls) continue;
-        out->is_valid = false;
-        return;
-      }
-      if (!valid) {
-        value = UnboxScalar<OutType>::Unbox(scalar);
-        valid = true;
-      } else {
-        value = Op::Call(value, UnboxScalar<OutType>::Unbox(scalar));
-      }
-    }
-    out->is_valid = valid;
-    if (valid) {
-      BoxScalar<OutType>::Box(value, out);
-    }
-  }
-
-  static Status Exec(KernelContext* ctx, const ExecBatch& batch, Datum* out) {
-    const ElementWiseAggregateOptions& options = MinMaxState::Get(ctx);
-    const auto descrs = batch.GetDescriptors();
-    const size_t scalar_count =
-        static_cast<size_t>(std::count_if(batch.values.begin(), batch.values.end(),
-                                          [](const Datum& d) { return d.is_scalar(); }));
-    if (scalar_count == batch.values.size()) {
-      ExecScalar(batch, options, out->scalar().get());
-      return Status::OK();
-    }
-
-    ArrayData* output = out->mutable_array();
-
-    // At least one array, two or more arguments
-    ArrayDataVector arrays;
-    for (const auto& arg : batch.values) {
-      if (!arg.is_array()) continue;
-      arrays.push_back(arg.array());
-    }
-
-    bool initialize_output = true;
-    if (scalar_count > 0) {
-      ARROW_ASSIGN_OR_RAISE(std::shared_ptr<Scalar> temp_scalar,
-                            MakeScalar(out->type(), 0));
-      ExecScalar(batch, options, temp_scalar.get());
-      if (temp_scalar->is_valid) {
-        const auto value = UnboxScalar<OutType>::Unbox(*temp_scalar);
-        initialize_output = false;
-        OutValue* out = output->GetMutableValues<OutValue>(1);
-        std::fill(out, out + batch.length, value);
-      } else if (!options.skip_nulls) {
-        // Abort early
-        ARROW_ASSIGN_OR_RAISE(auto array, MakeArrayFromScalar(*temp_scalar, batch.length,
-                                                              ctx->memory_pool()));
-        *output = *array->data();
-        return Status::OK();
-      }
-    }
-
-    if (initialize_output) {
-      OutValue* out = output->GetMutableValues<OutValue>(1);
-      std::fill(out, out + batch.length, Op::template antiextreme<OutValue>());
-    }
-
-    // Precompute the validity buffer
-    if (options.skip_nulls && initialize_output) {
-      // OR together the validity buffers of all arrays
-      if (std::all_of(arrays.begin(), arrays.end(),
-                      [](const std::shared_ptr<ArrayData>& arr) {
-                        return arr->MayHaveNulls();
-                      })) {
-        for (const auto& arr : arrays) {
-          if (!arr->MayHaveNulls()) continue;
-          if (!output->buffers[0]) {
-            ARROW_ASSIGN_OR_RAISE(output->buffers[0], ctx->AllocateBitmap(batch.length));
-            ::arrow::internal::CopyBitmap(arr->buffers[0]->data(), arr->offset,
-                                          batch.length,
-                                          output->buffers[0]->mutable_data(),
-                                          /*dest_offset=*/0);
-          } else {
-            ::arrow::internal::BitmapOr(
-                output->buffers[0]->data(), /*left_offset=*/0, arr->buffers[0]->data(),
-                arr->offset, batch.length,
-                /*out_offset=*/0, output->buffers[0]->mutable_data());
-          }
-        }
-      }
-    } else if (!options.skip_nulls) {
-      // AND together the validity buffers of all arrays
-      for (const auto& arr : arrays) {
-        if (!arr->MayHaveNulls()) continue;
-        if (!output->buffers[0]) {
-          ARROW_ASSIGN_OR_RAISE(output->buffers[0], ctx->AllocateBitmap(batch.length));
-          ::arrow::internal::CopyBitmap(arr->buffers[0]->data(), arr->offset,
-                                        batch.length, output->buffers[0]->mutable_data(),
-                                        /*dest_offset=*/0);
-        } else {
-          ::arrow::internal::BitmapAnd(output->buffers[0]->data(), /*left_offset=*/0,
-                                       arr->buffers[0]->data(), arr->offset, batch.length,
-                                       /*out_offset=*/0,
-                                       output->buffers[0]->mutable_data());
-        }
-      }
-    }
-
-    for (const auto& array : arrays) {
-      OutputArrayWriter<OutType> writer(out->mutable_array());
-      ArrayIterator<OutType> out_it(*output);
-      int64_t index = 0;
-      VisitArrayValuesInline<OutType>(
-          *array,
-          [&](OutValue value) {
-            auto u = out_it();
-            if (!output->buffers[0] ||
-                BitUtil::GetBit(output->buffers[0]->data(), index)) {
-              writer.Write(Op::Call(u, value));
-            } else {
-              writer.Write(value);
-            }
-            index++;
-          },
-          [&]() {
-            // RHS is null, preserve the LHS
-            writer.values++;
-            index++;
-            out_it();
-          });
-    }
-    output->null_count = output->buffers[0] ? -1 : 0;
-    return Status::OK();
-  }
-};
-
-template <typename Op>
-std::shared_ptr<ScalarFunction> MakeScalarMinMax(std::string name,
-                                                 const FunctionDoc* doc) {
-  auto func = std::make_shared<ArithmeticVarArgsFunction>(name, Arity::VarArgs(), doc);
-  for (const auto& ty : NumericTypes()) {
-    auto exec = GeneratePhysicalNumeric<ScalarMinMax, Op>(ty);
-    ScalarKernel kernel{KernelSignature::Make({ty}, ty, /*is_varargs=*/true), exec,
-                        MinMaxState::Init};
-    kernel.null_handling = NullHandling::type::COMPUTED_NO_PREALLOCATE;
-    kernel.mem_allocation = MemAllocation::type::PREALLOCATE;
-    DCHECK_OK(func->AddKernel(std::move(kernel)));
-  }
-  for (const auto& ty : TemporalTypes()) {
-    auto exec = GeneratePhysicalNumeric<ScalarMinMax, Op>(ty);
-    ScalarKernel kernel{KernelSignature::Make({ty}, ty, /*is_varargs=*/true), exec,
-                        MinMaxState::Init};
-    kernel.null_handling = NullHandling::type::COMPUTED_NO_PREALLOCATE;
-    kernel.mem_allocation = MemAllocation::type::PREALLOCATE;
-    DCHECK_OK(func->AddKernel(std::move(kernel)));
   }
   return func;
 }
@@ -850,20 +602,6 @@ const FunctionDoc pow_checked_doc{
     ("An error is returned when integer to negative integer power is encountered,\n"
      "or integer overflow is encountered."),
     {"base", "exponent"}};
-
-const FunctionDoc element_wise_min_doc{
-    "Find the element-wise minimum value",
-    ("Nulls will be ignored (default) or propagated. "
-     "NaN will be taken over null, but not over any valid float."),
-    {"*args"},
-    "ElementWiseAggregateOptions"};
-
-const FunctionDoc element_wise_max_doc{
-    "Find the element-wise maximum value",
-    ("Nulls will be ignored (default) or propagated. "
-     "NaN will be taken over null, but not over any valid float."),
-    {"*args"},
-    "ElementWiseAggregateOptions"};
 }  // namespace
 
 void RegisterScalarArithmetic(FunctionRegistry* registry) {
@@ -939,15 +677,6 @@ void RegisterScalarArithmetic(FunctionRegistry* registry) {
   auto power_checked =
       MakeArithmeticFunctionNotNull<PowerChecked>("power_checked", &pow_checked_doc);
   DCHECK_OK(registry->AddFunction(std::move(power_checked)));
-
-  // ----------------------------------------------------------------------
-  auto element_wise_min =
-      MakeScalarMinMax<Minimum>("element_wise_min", &element_wise_min_doc);
-  DCHECK_OK(registry->AddFunction(std::move(element_wise_min)));
-
-  auto element_wise_max =
-      MakeScalarMinMax<Maximum>("element_wise_max", &element_wise_max_doc);
-  DCHECK_OK(registry->AddFunction(std::move(element_wise_max)));
 }
 
 }  // namespace internal
