+++ cpp/src/arrow/compute/kernels/scalar_arithmetic_test.cc
@@ -312,74 +312,6 @@ class TestBinaryArithmeticUnsigned : public TestBinaryArithmeticIntegral<T> {};
 template <typename T>
 class TestBinaryArithmeticFloating : public TestBinaryArithmetic<T> {};
 
-template <typename T>
-class TestVarArgsArithmetic : public TestBase {
- protected:
-  static std::shared_ptr<DataType> type_singleton() {
-    return TypeTraits<T>::type_singleton();
-  }
-
-  using VarArgsFunction = std::function<Result<Datum>(
-      const std::vector<Datum>&, ElementWiseAggregateOptions, ExecContext*)>;
-
-  void SetUp() override { equal_options_ = equal_options_.nans_equal(true); }
-
-  Datum scalar(const std::string& value) {
-    return ScalarFromJSON(type_singleton(), value);
-  }
-
-  Datum array(const std::string& value) { return ArrayFromJSON(type_singleton(), value); }
-
-  Datum Eval(VarArgsFunction func, const std::vector<Datum>& args) {
-    EXPECT_OK_AND_ASSIGN(auto actual,
-                         func(args, element_wise_aggregate_options_, nullptr));
-    if (actual.is_array()) {
-      auto arr = actual.make_array();
-      ARROW_EXPECT_OK(arr->ValidateFull());
-    }
-    return actual;
-  }
-
-  void AssertNullScalar(VarArgsFunction func, const std::vector<Datum>& args) {
-    auto datum = this->Eval(func, args);
-    ASSERT_TRUE(datum.is_scalar());
-    ASSERT_FALSE(datum.scalar()->is_valid);
-  }
-
-  void Assert(VarArgsFunction func, Datum expected, const std::vector<Datum>& args) {
-    auto actual = Eval(func, args);
-    AssertDatumsApproxEqual(expected, actual, /*verbose=*/true, equal_options_);
-  }
-
-  EqualOptions equal_options_ = EqualOptions::Defaults();
-  ElementWiseAggregateOptions element_wise_aggregate_options_;
-};
-
-template <typename T>
-class TestVarArgsArithmeticNumeric : public TestVarArgsArithmetic<T> {};
-
-template <typename T>
-class TestVarArgsArithmeticFloating : public TestVarArgsArithmetic<T> {};
-
-template <typename T>
-class TestVarArgsArithmeticParametricTemporal : public TestVarArgsArithmetic<T> {
- protected:
-  static std::shared_ptr<DataType> type_singleton() {
-    // Time32 requires second/milli, Time64 requires nano/micro
-    if (TypeTraits<T>::bytes_required(1) == 4) {
-      return std::make_shared<T>(TimeUnit::type::SECOND);
-    } else {
-      return std::make_shared<T>(TimeUnit::type::NANO);
-    }
-  }
-
-  Datum scalar(const std::string& value) {
-    return ScalarFromJSON(type_singleton(), value);
-  }
-
-  Datum array(const std::string& value) { return ArrayFromJSON(type_singleton(), value); }
-};
-
 // InputType - OutputType pairs
 using IntegralTypes = testing::Types<Int8Type, Int16Type, Int32Type, Int64Type, UInt8Type,
                                      UInt16Type, UInt32Type, UInt64Type>;
@@ -392,12 +324,6 @@ using UnsignedIntegerTypes =
 // TODO(kszucs): add half-float
 using FloatingTypes = testing::Types<FloatType, DoubleType>;
 
-using NumericBasedTypes =
-    ::testing::Types<UInt8Type, UInt16Type, UInt32Type, UInt64Type, Int8Type, Int16Type,
-                     Int32Type, Int64Type, FloatType, DoubleType, Date32Type, Date64Type>;
-
-using ParametricTemporalTypes = ::testing::Types<TimestampType, Time32Type, Time64Type>;
-
 TYPED_TEST_SUITE(TestUnaryArithmeticIntegral, IntegralTypes);
 TYPED_TEST_SUITE(TestUnaryArithmeticSigned, SignedIntegerTypes);
 TYPED_TEST_SUITE(TestUnaryArithmeticUnsigned, UnsignedIntegerTypes);
@@ -407,10 +333,6 @@ TYPED_TEST_SUITE(TestBinaryArithmeticIntegral, IntegralTypes);
 TYPED_TEST_SUITE(TestBinaryArithmeticSigned, SignedIntegerTypes);
 TYPED_TEST_SUITE(TestBinaryArithmeticUnsigned, UnsignedIntegerTypes);
 TYPED_TEST_SUITE(TestBinaryArithmeticFloating, FloatingTypes);
-
-TYPED_TEST_SUITE(TestVarArgsArithmeticNumeric, NumericBasedTypes);
-TYPED_TEST_SUITE(TestVarArgsArithmeticFloating, FloatingTypes);
-TYPED_TEST_SUITE(TestVarArgsArithmeticParametricTemporal, ParametricTemporalTypes);
 
 TYPED_TEST(TestBinaryArithmeticIntegral, Add) {
   for (auto check_overflow : {false, true}) {
@@ -1236,277 +1158,6 @@ TYPED_TEST(TestUnaryArithmeticFloating, AbsoluteValue) {
     // Min/max
     this->AssertUnaryOp(AbsoluteValue, min, max);
     this->AssertUnaryOp(AbsoluteValue, max, max);
-  }
-}
-
-TYPED_TEST(TestVarArgsArithmeticNumeric, ElementWiseMin) {
-  this->AssertNullScalar(ElementWiseMin, {});
-  this->AssertNullScalar(ElementWiseMin, {this->scalar("null"), this->scalar("null")});
-
-  this->Assert(ElementWiseMin, this->scalar("0"), {this->scalar("0")});
-  this->Assert(ElementWiseMin, this->scalar("0"),
-               {this->scalar("2"), this->scalar("0"), this->scalar("1")});
-  this->Assert(
-      ElementWiseMin, this->scalar("0"),
-      {this->scalar("2"), this->scalar("0"), this->scalar("1"), this->scalar("null")});
-  this->Assert(ElementWiseMin, this->scalar("1"),
-               {this->scalar("null"), this->scalar("null"), this->scalar("1"),
-                this->scalar("null")});
-
-  this->Assert(ElementWiseMin, (this->array("[]")), {this->array("[]")});
-  this->Assert(ElementWiseMin, this->array("[1, 2, 3, null]"),
-               {this->array("[1, 2, 3, null]")});
-
-  this->Assert(ElementWiseMin, this->array("[1, 2, 2, 2]"),
-               {this->array("[1, 2, 3, 4]"), this->scalar("2")});
-  this->Assert(ElementWiseMin, this->array("[1, 2, 2, 2]"),
-               {this->array("[1, null, 3, 4]"), this->scalar("2")});
-  this->Assert(ElementWiseMin, this->array("[1, 2, 2, 2]"),
-               {this->array("[1, null, 3, 4]"), this->scalar("2"), this->scalar("4")});
-  this->Assert(ElementWiseMin, this->array("[1, 2, 2, 2]"),
-               {this->array("[1, null, 3, 4]"), this->scalar("null"), this->scalar("2")});
-
-  this->Assert(ElementWiseMin, this->array("[1, 2, 2, 2]"),
-               {this->array("[1, 2, 3, 4]"), this->array("[2, 2, 2, 2]")});
-  this->Assert(ElementWiseMin, this->array("[1, 2, 2, 2]"),
-               {this->array("[1, 2, 3, 4]"), this->array("[2, null, 2, 2]")});
-  this->Assert(ElementWiseMin, this->array("[1, 2, 2, 2]"),
-               {this->array("[1, null, 3, 4]"), this->array("[2, 2, 2, 2]")});
-
-  this->Assert(ElementWiseMin, this->array("[1, 2, null, 6]"),
-               {this->array("[1, 2, null, null]"), this->array("[4, null, null, 6]")});
-  this->Assert(ElementWiseMin, this->array("[1, 2, null, 6]"),
-               {this->array("[4, null, null, 6]"), this->array("[1, 2, null, null]")});
-  this->Assert(ElementWiseMin, this->array("[1, 2, 3, 4]"),
-               {this->array("[1, 2, 3, 4]"), this->array("[null, null, null, null]")});
-  this->Assert(ElementWiseMin, this->array("[1, 2, 3, 4]"),
-               {this->array("[null, null, null, null]"), this->array("[1, 2, 3, 4]")});
-
-  this->Assert(ElementWiseMin, this->array("[1, 1, 1, 1]"),
-               {this->scalar("1"), this->array("[1, 2, 3, 4]")});
-  this->Assert(ElementWiseMin, this->array("[1, 1, 1, 1]"),
-               {this->scalar("1"), this->array("[null, null, null, null]")});
-  this->Assert(ElementWiseMin, this->array("[1, 1, 1, 1]"),
-               {this->scalar("null"), this->array("[1, 1, 1, 1]")});
-  this->Assert(ElementWiseMin, this->array("[null, null, null, null]"),
-               {this->scalar("null"), this->array("[null, null, null, null]")});
-
-  // Test null handling
-  this->element_wise_aggregate_options_.skip_nulls = false;
-  this->AssertNullScalar(ElementWiseMin, {this->scalar("null"), this->scalar("null")});
-  this->AssertNullScalar(ElementWiseMin, {this->scalar("0"), this->scalar("null")});
-
-  this->Assert(ElementWiseMin, this->array("[1, null, 2, 2]"),
-               {this->array("[1, null, 3, 4]"), this->scalar("2"), this->scalar("4")});
-  this->Assert(ElementWiseMin, this->array("[null, null, null, null]"),
-               {this->array("[1, null, 3, 4]"), this->scalar("null"), this->scalar("2")});
-  this->Assert(ElementWiseMin, this->array("[1, null, 2, 2]"),
-               {this->array("[1, 2, 3, 4]"), this->array("[2, null, 2, 2]")});
-
-  this->Assert(ElementWiseMin, this->array("[null, null, null, null]"),
-               {this->scalar("1"), this->array("[null, null, null, null]")});
-  this->Assert(ElementWiseMin, this->array("[null, null, null, null]"),
-               {this->scalar("null"), this->array("[1, 1, 1, 1]")});
-}
-
-TYPED_TEST(TestVarArgsArithmeticFloating, ElementWiseMin) {
-  auto Check = [this](const std::string& expected,
-                      const std::vector<std::string>& inputs) {
-    std::vector<Datum> args;
-    for (const auto& input : inputs) {
-      args.emplace_back(this->scalar(input));
-    }
-    this->Assert(ElementWiseMin, this->scalar(expected), args);
-
-    args.clear();
-    for (const auto& input : inputs) {
-      args.emplace_back(this->array("[" + input + "]"));
-    }
-    this->Assert(ElementWiseMin, this->array("[" + expected + "]"), args);
-  };
-  Check("-0.0", {"0.0", "-0.0"});
-  Check("-0.0", {"1.0", "-0.0", "0.0"});
-  Check("-1.0", {"-1.0", "-0.0"});
-  Check("0", {"0", "NaN"});
-  Check("0", {"NaN", "0"});
-  Check("Inf", {"Inf", "NaN"});
-  Check("Inf", {"NaN", "Inf"});
-  Check("-Inf", {"-Inf", "NaN"});
-  Check("-Inf", {"NaN", "-Inf"});
-  Check("NaN", {"NaN", "null"});
-  Check("0", {"0", "Inf"});
-  Check("-Inf", {"0", "-Inf"});
-}
-
-TYPED_TEST(TestVarArgsArithmeticParametricTemporal, ElementWiseMin) {
-  // Temporal kernel is implemented with numeric kernel underneath
-  this->AssertNullScalar(ElementWiseMin, {});
-  this->AssertNullScalar(ElementWiseMin, {this->scalar("null"), this->scalar("null")});
-
-  this->Assert(ElementWiseMin, this->scalar("0"), {this->scalar("0")});
-  this->Assert(ElementWiseMin, this->scalar("0"), {this->scalar("2"), this->scalar("0")});
-  this->Assert(ElementWiseMin, this->scalar("0"),
-               {this->scalar("0"), this->scalar("null")});
-
-  this->Assert(ElementWiseMin, (this->array("[]")), {this->array("[]")});
-  this->Assert(ElementWiseMin, this->array("[1, 2, 3, null]"),
-               {this->array("[1, 2, 3, null]")});
-
-  this->Assert(ElementWiseMin, this->array("[1, 2, 2, 2]"),
-               {this->array("[1, null, 3, 4]"), this->scalar("null"), this->scalar("2")});
-
-  this->Assert(ElementWiseMin, this->array("[1, 2, 3, 2]"),
-               {this->array("[1, null, 3, 4]"), this->array("[2, 2, null, 2]")});
-}
-
-TYPED_TEST(TestVarArgsArithmeticNumeric, ElementWiseMax) {
-  this->AssertNullScalar(ElementWiseMax, {});
-  this->AssertNullScalar(ElementWiseMax, {this->scalar("null"), this->scalar("null")});
-
-  this->Assert(ElementWiseMax, this->scalar("0"), {this->scalar("0")});
-  this->Assert(ElementWiseMax, this->scalar("2"),
-               {this->scalar("2"), this->scalar("0"), this->scalar("1")});
-  this->Assert(
-      ElementWiseMax, this->scalar("2"),
-      {this->scalar("2"), this->scalar("0"), this->scalar("1"), this->scalar("null")});
-  this->Assert(ElementWiseMax, this->scalar("1"),
-               {this->scalar("null"), this->scalar("null"), this->scalar("1"),
-                this->scalar("null")});
-
-  this->Assert(ElementWiseMax, (this->array("[]")), {this->array("[]")});
-  this->Assert(ElementWiseMax, this->array("[1, 2, 3, null]"),
-               {this->array("[1, 2, 3, null]")});
-
-  this->Assert(ElementWiseMax, this->array("[2, 2, 3, 4]"),
-               {this->array("[1, 2, 3, 4]"), this->scalar("2")});
-  this->Assert(ElementWiseMax, this->array("[2, 2, 3, 4]"),
-               {this->array("[1, null, 3, 4]"), this->scalar("2")});
-  this->Assert(ElementWiseMax, this->array("[4, 4, 4, 4]"),
-               {this->array("[1, null, 3, 4]"), this->scalar("2"), this->scalar("4")});
-  this->Assert(ElementWiseMax, this->array("[2, 2, 3, 4]"),
-               {this->array("[1, null, 3, 4]"), this->scalar("null"), this->scalar("2")});
-
-  this->Assert(ElementWiseMax, this->array("[2, 2, 3, 4]"),
-               {this->array("[1, 2, 3, 4]"), this->array("[2, 2, 2, 2]")});
-  this->Assert(ElementWiseMax, this->array("[2, 2, 3, 4]"),
-               {this->array("[1, 2, 3, 4]"), this->array("[2, null, 2, 2]")});
-  this->Assert(ElementWiseMax, this->array("[2, 2, 3, 4]"),
-               {this->array("[1, null, 3, 4]"), this->array("[2, 2, 2, 2]")});
-
-  this->Assert(ElementWiseMax, this->array("[4, 2, null, 6]"),
-               {this->array("[1, 2, null, null]"), this->array("[4, null, null, 6]")});
-  this->Assert(ElementWiseMax, this->array("[4, 2, null, 6]"),
-               {this->array("[4, null, null, 6]"), this->array("[1, 2, null, null]")});
-  this->Assert(ElementWiseMax, this->array("[1, 2, 3, 4]"),
-               {this->array("[1, 2, 3, 4]"), this->array("[null, null, null, null]")});
-  this->Assert(ElementWiseMax, this->array("[1, 2, 3, 4]"),
-               {this->array("[null, null, null, null]"), this->array("[1, 2, 3, 4]")});
-
-  this->Assert(ElementWiseMax, this->array("[1, 2, 3, 4]"),
-               {this->scalar("1"), this->array("[1, 2, 3, 4]")});
-  this->Assert(ElementWiseMax, this->array("[1, 1, 1, 1]"),
-               {this->scalar("1"), this->array("[null, null, null, null]")});
-  this->Assert(ElementWiseMax, this->array("[1, 1, 1, 1]"),
-               {this->scalar("null"), this->array("[1, 1, 1, 1]")});
-  this->Assert(ElementWiseMax, this->array("[null, null, null, null]"),
-               {this->scalar("null"), this->array("[null, null, null, null]")});
-
-  // Test null handling
-  this->element_wise_aggregate_options_.skip_nulls = false;
-  this->AssertNullScalar(ElementWiseMax, {this->scalar("null"), this->scalar("null")});
-  this->AssertNullScalar(ElementWiseMax, {this->scalar("0"), this->scalar("null")});
-
-  this->Assert(ElementWiseMax, this->array("[4, null, 4, 4]"),
-               {this->array("[1, null, 3, 4]"), this->scalar("2"), this->scalar("4")});
-  this->Assert(ElementWiseMax, this->array("[null, null, null, null]"),
-               {this->array("[1, null, 3, 4]"), this->scalar("null"), this->scalar("2")});
-  this->Assert(ElementWiseMax, this->array("[2, null, 3, 4]"),
-               {this->array("[1, 2, 3, 4]"), this->array("[2, null, 2, 2]")});
-
-  this->Assert(ElementWiseMax, this->array("[null, null, null, null]"),
-               {this->scalar("1"), this->array("[null, null, null, null]")});
-  this->Assert(ElementWiseMax, this->array("[null, null, null, null]"),
-               {this->scalar("null"), this->array("[1, 1, 1, 1]")});
-}
-
-TYPED_TEST(TestVarArgsArithmeticFloating, ElementWiseMax) {
-  auto Check = [this](const std::string& expected,
-                      const std::vector<std::string>& inputs) {
-    std::vector<Datum> args;
-    for (const auto& input : inputs) {
-      args.emplace_back(this->scalar(input));
-    }
-    this->Assert(ElementWiseMax, this->scalar(expected), args);
-
-    args.clear();
-    for (const auto& input : inputs) {
-      args.emplace_back(this->array("[" + input + "]"));
-    }
-    this->Assert(ElementWiseMax, this->array("[" + expected + "]"), args);
-  };
-  Check("0.0", {"0.0", "-0.0"});
-  Check("1.0", {"1.0", "-0.0", "0.0"});
-  Check("-0.0", {"-1.0", "-0.0"});
-  Check("0", {"0", "NaN"});
-  Check("0", {"NaN", "0"});
-  Check("Inf", {"Inf", "NaN"});
-  Check("Inf", {"NaN", "Inf"});
-  Check("-Inf", {"-Inf", "NaN"});
-  Check("-Inf", {"NaN", "-Inf"});
-  Check("NaN", {"NaN", "null"});
-  Check("Inf", {"0", "Inf"});
-  Check("0", {"0", "-Inf"});
-}
-
-TYPED_TEST(TestVarArgsArithmeticParametricTemporal, ElementWiseMax) {
-  // Temporal kernel is implemented with numeric kernel underneath
-  this->AssertNullScalar(ElementWiseMax, {});
-  this->AssertNullScalar(ElementWiseMax, {this->scalar("null"), this->scalar("null")});
-
-  this->Assert(ElementWiseMax, this->scalar("0"), {this->scalar("0")});
-  this->Assert(ElementWiseMax, this->scalar("2"), {this->scalar("2"), this->scalar("0")});
-  this->Assert(ElementWiseMax, this->scalar("0"),
-               {this->scalar("0"), this->scalar("null")});
-
-  this->Assert(ElementWiseMax, (this->array("[]")), {this->array("[]")});
-  this->Assert(ElementWiseMax, this->array("[1, 2, 3, null]"),
-               {this->array("[1, 2, 3, null]")});
-
-  this->Assert(ElementWiseMax, this->array("[2, 2, 3, 4]"),
-               {this->array("[1, null, 3, 4]"), this->scalar("null"), this->scalar("2")});
-
-  this->Assert(ElementWiseMax, this->array("[2, 2, 3, 4]"),
-               {this->array("[1, null, 3, 4]"), this->array("[2, 2, null, 2]")});
-}
-
-TEST(TestElementWiseMaxElementWiseMin, CommonTimestamp) {
-  {
-    auto t1 = std::make_shared<TimestampType>(TimeUnit::SECOND);
-    auto t2 = std::make_shared<TimestampType>(TimeUnit::MILLI);
-    auto expected = MakeScalar(t2, 1000).ValueOrDie();
-    ASSERT_OK_AND_ASSIGN(auto actual,
-                         ElementWiseMin({Datum(MakeScalar(t1, 1).ValueOrDie()),
-                                         Datum(MakeScalar(t2, 12000).ValueOrDie())}));
-    AssertScalarsEqual(*expected, *actual.scalar(), /*verbose=*/true);
-  }
-  {
-    auto t1 = std::make_shared<Date32Type>();
-    auto t2 = std::make_shared<TimestampType>(TimeUnit::SECOND);
-    auto expected = MakeScalar(t2, 86401).ValueOrDie();
-    ASSERT_OK_AND_ASSIGN(auto actual,
-                         ElementWiseMax({Datum(MakeScalar(t1, 1).ValueOrDie()),
-                                         Datum(MakeScalar(t2, 86401).ValueOrDie())}));
-    AssertScalarsEqual(*expected, *actual.scalar(), /*verbose=*/true);
-  }
-  {
-    auto t1 = std::make_shared<Date32Type>();
-    auto t2 = std::make_shared<Date64Type>();
-    auto t3 = std::make_shared<TimestampType>(TimeUnit::SECOND);
-    auto expected = MakeScalar(t3, 86400).ValueOrDie();
-    ASSERT_OK_AND_ASSIGN(
-        auto actual, ElementWiseMin({Datum(MakeScalar(t1, 1).ValueOrDie()),
-                                     Datum(MakeScalar(t2, 2 * 86400000).ValueOrDie())}));
-    AssertScalarsEqual(*expected, *actual.scalar(), /*verbose=*/true);
   }
 }
